// src/app/models/masters-mpc/classification.ts
export class classification {
  id: number=-1;
  name: string= "";
  createdByUser:string ="";
  createdByUserId:number=-1;
  active: boolean = false;
  indInitialConfiguration: boolean=false;
  abbreviation: string = "";
  internalCode: string = "";
  updatedByUser:string="";
  createdDate: Date;
}